requirements.txt
flask
flask_restful
sqlalchemy
pandas
python-dotenv
mysql-connector